refactor: move all handlers in their own file
